restructure project stucture
