+++ fire/core/algorithm/_numba_funcs.py
