+++ fire/evaluation/industry/evaluator.py
@@ -5,8 +5,8 @@ import typing
 
 import pandas as pd
 
-from . import plots
-from .eva_utils import ForwardReturns, IC, QuantileReturns, compute_ic, compute_quantile_returns
+from ...core.plot import plots
+from ..eva_utils import ForwardReturns, IC, QuantileReturns, compute_ic, compute_quantile_returns
 
 __all__ = ["Evaluator"]
